@@ -1,10 +1,4 @@
 package org.acme.experiments.dto;
 
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class StatusDTO {
-   public boolean verified;
-   public int sentCount;
-   public String feedback;
+public record StatusDTO (boolean verified, int sentCount, String feedback) {
 }
